+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 - Mouse wheel can be used to adjust volume while hovering over the volume control (slow by default, faster with SHIFT)
+- Button to dismiss the "files to scan" banner without taking any action (it will be shown again on page reload)
+  [#1251](https://github.com/owncloud/music/issues/1251)
 - Belarusian translation
   [#1254](https://github.com/owncloud/music/pull/1254) @pavel-miniutka